initial implementation of limiting Newton increment to satisfy bounds     e.g. tracer > 0
For each tracer module, compute scalef to ensure iterate + scalef * increment
is within bounds.

Tracer bounds are obtained from tracer_module_defs.yaml.
    This can be overridden by overriding TracerModuleStateBase.get_bounds in
    specific tracer modules.

This implementation applies a single scalef to all tracers in a tracer module.
In particular, scalef is independent of region. This should be generalized.
An issue will be filed on this limitation.

Changes to be committed:
	modified:   src/model_state_base.py
	modified:   src/newton_solver.py
	modified:   src/tracer_module_state_base.py
	modified:   src/utils.py